New RPC for manual removal of InstantSend lock (#1680)
A situation where old IS-locked transactions for some reason do not get mined, for e.g. a mining pool clears their mempool, results in these IS-locked transactions never getting mined yet remaining in the IS-lock database on some nodes. If a conflicting transaction emerges, for example if the original sender of the initial IS-locked tries to use the input again, it then results in a conflict. This fix is part of a larger fix that will be implemented later.